backend/requirements/common.txt
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-framework==0.1.0
djangorestframework==3.12.1
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.3
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
traitlets==5.0.4
wcwidth==0.2.5
wrapt==1.12.1
